move the slice utils into a separate module
